refactor: simplify SQL builder dispatch
