+++ src/domain/mysql_sql/mod.rs
@@ -494,6 +494,8 @@ fn consume_required_ascii_whitespace(text: &str, index: &mut usize) -> Option<()
 mod tests {
     use super::*;
 
+    include!("tests.rs");
+
     #[test]
     fn splits_mysql_comments_quotes_and_backticks() {
         let sql = "SELECT 'a;\\'b'; # comment;\n SELECT `semi;colon` /* ; */";
