+++ src/domain/mysql_sql/sql.rs
@@ -120,61 +120,3 @@ fn quote_string(value: &str) -> String {
     }
     format!("'{escaped}'")
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn builds_tree_explain_sql() {
-        assert_eq!(
-            build_explain_sql("SELECT * FROM users"),
-            Some("EXPLAIN FORMAT=TREE SELECT * FROM users".to_string())
-        );
-        assert_eq!(build_explain_sql("SELECT 1; SELECT 2"), None);
-        assert_eq!(
-            build_explain_analyze_sql("SELECT * FROM users"),
-            Some("EXPLAIN ANALYZE FORMAT=TREE SELECT * FROM users".to_string())
-        );
-        assert_eq!(
-            build_explain_analyze_sql("UPDATE users SET name = 'Ada'"),
-            None
-        );
-    }
-
-    #[test]
-    fn builds_update_sql_with_mysql_literals() {
-        assert_eq!(
-            build_update_sql(
-                "db`name",
-                "table`name",
-                "value`name",
-                &QueryValue::text("O'Reilly\\path\n\t\0"),
-                &[
-                    ("id`part".into(), QueryValue::SqlLiteral("1".into())),
-                    ("tenant".into(), QueryValue::Null),
-                ],
-            ),
-            "UPDATE `db``name`.`table``name`\nSET `value``name` = 'O\\'Reilly\\\\path\\n\\t\\0'\nWHERE `id``part` = 1 AND `tenant` IS NULL;"
-        );
-    }
-
-    #[test]
-    fn builds_bulk_delete_sql_with_row_grouping() {
-        assert_eq!(
-            build_bulk_delete_sql(
-                "sabiql_test",
-                "items",
-                &[
-                    vec![("first".into(), QueryValue::SqlLiteral("1".into()))],
-                    vec![("first".into(), QueryValue::SqlLiteral("2".into()))],
-                ],
-            ),
-            "DELETE FROM `sabiql_test`.`items`\nWHERE (`first` = 1) OR (`first` = 2);"
-        );
-    }
-}
-
-#[cfg(test)]
-#[path = "../mysql_sql_legacy_tests.rs"]
-mod legacy_tests;
