+++ src/domain/postgres_sql.rs
@@ -102,71 +102,5 @@ fn row_predicate(pk_pairs: &[(String, QueryValue)]) -> String {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
-    #[test]
-    fn builds_explain_sql() {
-        assert_eq!(
-            build_explain_sql("SELECT 1"),
-            Some("EXPLAIN SELECT 1".to_string())
-        );
-        assert_eq!(
-            build_explain_analyze_sql("SELECT 1"),
-            Some("EXPLAIN ANALYZE SELECT 1".to_string())
-        );
-    }
-
-    #[test]
-    fn builds_update_sql_with_postgres_literals() {
-        assert_eq!(
-            build_update_sql(
-                "public",
-                "users",
-                "name",
-                &QueryValue::text("O'Reilly"),
-                &[("id".into(), QueryValue::Null)],
-            ),
-            "UPDATE \"public\".\"users\"\nSET \"name\" = 'O''Reilly'\nWHERE \"id\" IS NULL;"
-        );
-        assert_eq!(
-            build_update_sql(
-                "s",
-                "t",
-                "payload",
-                &QueryValue::Blob(vec![0, 255]),
-                &[("id\"part".into(), QueryValue::SqlLiteral("1".into()))],
-            ),
-            "UPDATE \"s\".\"t\"\nSET \"payload\" = '\\x00ff'\nWHERE \"id\"\"part\" = 1;"
-        );
-    }
-
-    #[test]
-    fn builds_bulk_delete_sql_with_row_grouping() {
-        assert_eq!(
-            build_bulk_delete_sql(
-                "public",
-                "users",
-                &[
-                    vec![("id".into(), QueryValue::text("1"))],
-                    vec![("id".into(), QueryValue::text("2"))],
-                ],
-            ),
-            "DELETE FROM \"public\".\"users\"\nWHERE (\"id\" = '1') OR (\"id\" = '2');"
-        );
-        assert_eq!(
-            build_bulk_delete_sql(
-                "s",
-                "t",
-                &[vec![
-                    ("id".into(), QueryValue::Null),
-                    ("tenant_id".into(), QueryValue::text("a")),
-                ]],
-            ),
-            "DELETE FROM \"s\".\"t\"\nWHERE \"id\" IS NULL AND \"tenant_id\" = 'a';"
-        );
-    }
+    include!("postgres_sql_tests.rs");
 }
-
-#[cfg(test)]
-#[path = "postgres_sql_legacy_tests.rs"]
-mod legacy_tests;
